backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
